# swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "Sample flask REST API",
    "version": "1.0.0",
    "title": "Arrested API",
    "termsOfService": "http://swagger.io/terms/",
    "contact": {
      "email": "ksmini82@gmail.com"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
  "host": "127.0.0.1:5000",
  "basePath": "/api/",
  "tags": [
    {
      "name": "company",
      "description": "회사 정보에 접근 할 수 있는 endpoint"
    },
    {
      "name": "tag",
      "description": "각 회사의 특징을 나타내는 태그 조회용 endpoint"
    }
  ],
  "schemes": [
    "http"
  ],
  "paths": {
    "/company": {
      "get": {
        "tags": [
          "company"
        ],
        "summary": "회사 목록 조회",
        "description": "",
        "operationId": "companyList",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "lang_code",
            "in": "query",
            "description": "언어 선택을 위한 언어코드",
            "required": false,
            "type": "string",
            "default": "ko"
          }
        ],
        "responses": {
          "200": {
            "schema": {
              "type": "object",
              "properties": {
                "num_results": {
                  "type": "integer",
                  "format": "int64",
                  "description": "전체 항목의 수"
                },
                "objects": {
                  "type": "array",
                  "items": {
                    "$ref": "#/definitions/company"
                  }
                },
                "page": {
                  "type": "integer",
                  "format": "int64",
                  "description": "조회된 페이지 번호"
                },
                "total_pages": {
                  "type": "integer",
                  "format": "int64",
                  "description": "전체 페이지 수"
                }
              }
            },
            "description": "OK"
          }
        }
      },
      "post": {
        "tags": [
          "company"
        ],
        "summary": "회사 추가",
        "description": "",
        "operationId": "addCompany",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "추가하려는 회사 정보",
            "required": true,
            "schema": {
              "type": "object",
              "required": [
                "name"
              ],
              "properties": {
                "name": {
                  "type": "string",
                  "example": "Arrested"
                },
                "lang_code": {
                  "type": "string",
                  "example": "ko",
                  "default": "ko"
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "추가 성공"
          },
          "400": {
            "description": "입력 데이터 오류"
          }
        }
      }
    },
    "/company/{company_id}/name": {
      "post": {
        "tags": [
          "company"
        ],
        "summary": "회사 정보에 다른 언어의 이름을 추가",
        "description": "",
        "operationId": "addNameToCompany",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "company_id",
            "in": "path",
            "description": "회사의 ID",
            "required": true,
            "type": "integer",
            "format": "int64"
          },
          {
            "in": "body",
            "name": "body",
            "description": "추가하려는 회사명을 입력",
            "required": true,
            "schema": {
              "type": "object",
              "required": [
                "company_name",
                "lang_code"
              ],
              "properties": {
                "company_name": {
                  "type": "string",
                  "example": "어레스티드"
                },
                "lang_code": {
                  "type": "string",
                  "example": "ko",
                  "default": "ko"
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "추가 성공"
          },
          "400": {
            "description": "입력 데이터 오류"
          }
        }
      },
      "delete": {
        "tags": [
          "company"
        ],
        "summary": "회사 정보에서 회사명을 제거",
        "description": "",
        "operationId": "removeNameFromCompany",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "company_id",
            "in": "path",
            "description": "회사의 ID",
            "required": true,
            "type": "integer",
            "format": "int64"
          },
          {
            "in": "body",
            "name": "body",
            "description": "삭제하려는 회사명을 입력",
            "required": true,
            "schema": {
              "type": "object",
              "required": [
                "company_name"
              ],
              "properties": {
                "company_name": {
                  "type": "string",
                  "example": "어레스티드"
                },
                "lang_code": {
                  "type": "string",
                  "example": "ko"
                }
              }
            }
          }
        ],
        "responses": {
          "204": {
            "description": "삭제 완료"
          },
          "400": {
            "description": "입력 데이터 오류"
          }
        }
      }
    },
    "/company/{company_id}/tag": {
      "post": {
        "tags": [
          "company"
        ],
        "summary": "회사 정보에 태그를 추가",
        "description": "",
        "operationId": "addTagToCompany",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "company_id",
            "in": "path",
            "description": "회사의 ID",
            "required": true,
            "type": "integer",
            "format": "int64"
          },
          {
            "in": "body",
            "name": "body",
            "description": "추가하려는 태그명을 입력",
            "required": true,
            "schema": {
              "type": "object",
              "required": [
                "tag_name"
              ],
              "properties": {
                "tag_name": {
                  "type": "string",
                  "example": "자율출근제"
                },
                "lang_code": {
                  "type": "string",
                  "example": "ko",
                  "default": "ko"
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "추가 성공"
          },
          "400": {
            "description": "입력 데이터 오류"
          }
        }
      },
      "delete": {
        "tags": [
          "company"
        ],
        "summary": "회사 정보에서 태그를 제거",
        "description": "",
        "operationId": "removeTagFromCompany",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "company_id",
            "in": "path",
            "description": "회사의 ID",
            "required": true,
            "type": "integer",
            "format": "int64"
          },
          {
            "in": "body",
            "name": "body",
            "description": "삭제하려는 태그명을 입력",
            "required": true,
            "schema": {
              "type": "object",
              "required": [
                "tag_name"
              ],
              "properties": {
                "tag_name": {
                  "type": "string",
                  "example": "자율출근제"
                }
              }
            }
          }
        ],
        "responses": {
          "204": {
            "description": "삭제 완료"
          },
          "400": {
            "description": "입력 데이터 오류"
          }
        }
      }
    },
    "/company/search_by_name": {
      "get": {
        "tags": [
          "company"
        ],
        "summary": "회사명으로 회사를 검색",
        "description": "",
        "operationId": "saerchCompanyByName",
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "company_name",
            "in": "query",
            "description": "회사명",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/company/search_by_tag": {
      "get": {
        "tags": [
          "company"
        ],
        "summary": "태그로 회사를 검색",
        "description": "",
        "operationId": "saerchCompanyByTag",
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "tag_name",
            "in": "query",
            "description": "태그명",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/tag": {
      "get": {
        "tags": [
          "tag"
        ],
        "summary": "전체 태크 목록 조회",
        "description": "",
        "operationId": "tagList",
        "produces": [
          "application/json"
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "object",
              "properties": {
                "num_results": {
                  "type": "integer",
                  "format": "int64",
                  "description": "전체 항목의 수"
                },
                "objects": {
                  "type": "array",
                  "items": {
                    "$ref": "#/definitions/tag"
                  }
                },
                "page": {
                  "type": "integer",
                  "format": "int64",
                  "description": "조회된 페이지 번호"
                },
                "total_pages": {
                  "type": "integer",
                  "format": "int64",
                  "description": "전체 페이지 수"
                }
              }
            }
          }
        }
      }
    },
    "/tag/{tag_id}/name": {
      "post": {
        "tags": [
          "tag"
        ],
        "summary": "태그에 다른 언어를 추가",
        "description": "",
        "operationId": "addNameToTag",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "tag_id",
            "in": "path",
            "description": "태그의 ID",
            "required": true,
            "type": "integer",
            "format": "int64"
          },
          {
            "in": "body",
            "name": "body",
            "description": "추가하려는 다른 언어의 태그명을 입력",
            "required": true,
            "schema": {
              "type": "object",
              "required": [
                "tag_name",
                "lang_code"
              ],
              "properties": {
                "tag_name": {
                  "type": "string",
                  "example": "Free tee time"
                },
                "lang_code": {
                  "type": "string",
                  "example": "en",
                  "default": "en"
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "추가 성공"
          },
          "400": {
            "description": "입력 데이터 오류"
          }
        }
      },
      "delete": {
        "tags": [
          "tag"
        ],
        "summary": "태그명 중 특정 언어를 제거",
        "description": "",
        "operationId": "removeNameFromTag",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "tag_id",
            "in": "path",
            "description": "태그의 ID",
            "required": true,
            "type": "integer",
            "format": "int64"
          },
          {
            "in": "body",
            "name": "body",
            "description": "삭제하려는 태그명을 입력",
            "required": true,
            "schema": {
              "type": "object",
              "required": [
                "tag_name",
                "lang_code"
              ],
              "properties": {
                "tag_name": {
                  "type": "string",
                  "example": "Free tee time"
                },
                "lang_code": {
                  "type": "string",
                  "example": "en"
                }
              }
            }
          }
        ],
        "responses": {
          "204": {
            "description": "삭제 완료"
          },
          "400": {
            "description": "입력 데이터 오류"
          }
        }
      }
    }
  },
  "definitions": {
    "company": {
      "type": "object",
      "required": [
        "name"
      ],
      "properties": {
        "id": {
          "type": "integer",
          "format": "int64"
        },
        "name": {
          "type": "string",
          "example": "Arrested"
        },
        "tags": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/tag"
          }
        },
        "created_at": {
          "type": "string"
        }
      }
    },
    "tag": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer",
          "format": "int64"
        },
        "name": {
          "type": "string"
        }
      }
    }
  }
}